- tested play_next_script_line_eeprom - tested play_next_script_line_progmem

+++ test/minunit.h
@@ -7,8 +7,8 @@
 	} while (0)
 
 #define mu_assert_eq(expected, actual) {\
-	int a = (expected); \
-	int b = (actual); \
+	uint8_t a = (expected); \
+	uint8_t b = (actual); \
 	do { if ((a) != (b)) { \
 		sprintf(minunit_msg, "ASSERT_EQ: %s:%d expected %d, got %u", __FILE__, __LINE__, a, b);\
 		return minunit_msg; }; \
